Add empirical detection of entropy of letters. Add normal approx of binomial distribution of edit similarity. Made fgclust.sh more user friendly to use. Changed Makefile to make executable for gprof performance profiling.

+++ bin/fgclust.sh
@@ -2,38 +2,41 @@
 #set -evx
 
 BINROOT=$(dirname `which $0`)
-if [ $# -lt 4 ]; then
-    echo "Usage: $0 <input-fasta-file> <output-fasta-file> <output-cluster-file> <sequence-similarity> <optional-params>"
-    echo "  input-fasta-file    : the input fasta, fastq, or gzipped-fastq file that contains either only nucleotide sequences or only protein sequences"
-    echo "  output-fasta-file   : the output fasta file that contains centroid sequences"
-    echo "  output-cluster-file : the output tsv file that contains the three following fields per line: "
-    echo "                        the representative centroid, the member covered by the centroid, and the sequence similarity between the centroid and the member"
-    echo "  sequence-similarity : the percent sequence similarity between 1 and 99"
-    echo "Optional parameters for sorting"
+if [ $# -lt 2 ]; then
+    echo "Usage: $0 <input-fasta-file> <output-fasta-file> <optional-parameters>"
+    echo "  <input-fasta-file>  : the input fasta, fastq, or gzipped-fastq file that contains either only nucleotide sequences or only protein sequences"
+    echo "  <output-fasta-file> : the output fasta file that contains centroid sequences"
+    echo "The <optional-parameters> for sorting"
     echo "  --israndom          : sort by decreasing sequence length instead of shuffling randomly"
     echo "Note: "
     echo "  the alphabet (which can be of either amino-acid or nucleotide) of the sequences in input-fasta-file is automatically detected"
-    echo "Optional parameters for clustering:"
-    "${BINROOT}/fastaseqs-to-distmatrix.out" --help
+    echo "  <output-fasta-file>.clu is the output tsv file that contains the three following fields per line: "
+    echo "                          the representative centroid, the member covered by the centroid, and the sequence similarity between the centroid and the member"
+    echo "  if <output-fasta-file> ends with the extension .keep, then all intermediate files are kept instead of bein removed by default"
+    echo "The <optional-parameters> for clustering:"
+    echo "" | "${BINROOT}/fastaseqs-to-distmatrix.out" --help
     exit 0
 fi
 
 INFILE="$1"
 OUTFASTA="$2"
-OUTCLUST="$3"
-SIM="$4"
+OUTCLUST="${2}.clu"
 
 OUTDIR="${OUTFASTA}_tmpdir"
 
-mkdir -p "${OUTDIR}" || true
+mkdir -p "${OUTDIR}/bin/" || true
+cp "${BINROOT}/"*.out "${OUTDIR}/bin/"
+BINROOT="${OUTDIR}/bin/"
 
-if [[ "$INFILE" == *.gz ]]; then catcmd=zcat; else catcmd=cat; fi
+if [[ "$INFILE" == *".gz" ]]; then catcmd=zcat; else catcmd=cat; fi
 
-$catcmd "${INFILE}"                 | "${BINROOT}/len-revname-sort.out"        "${@:5}"                     > "${OUTDIR}/shuf.fasta"
-cat     "${OUTDIR}/shuf.fasta"      | "${BINROOT}/fastaseqs-to-distmatrix.out" --sim-perc "${SIM}" "${@:5}" > "${OUTDIR}/distmatrix.tsv"
-cat     "${OUTDIR}/distmatrix.tsv"  | "${BINROOT}/linsetcover.out"                                          > "${OUTDIR}/ordsetcover.tsv"
-cat     "${OUTDIR}/ordsetcover.tsv" | "${BINROOT}/setcover-ords-to-hdrs.out"  "${OUTDIR}/shuf.fasta"        > "${OUTCLUST}"
-cat     "${OUTDIR}/ordsetcover.tsv" | "${BINROOT}/setcover-ords-to-fasta.out" "${OUTDIR}/shuf.fasta"        > "${OUTFASTA}"           
+$catcmd "${INFILE}"                 | "${BINROOT}/len-revname-sort.out"        "${@:3}"              > "${OUTDIR}/shuf.fasta"
+cat     "${OUTDIR}/shuf.fasta"      | "${BINROOT}/fastaseqs-to-distmatrix.out" "${@:3}"              > "${OUTDIR}/distmatrix.tsv"
+cat     "${OUTDIR}/distmatrix.tsv"  | "${BINROOT}/linsetcover.out"                                   > "${OUTDIR}/ordsetcover.tsv"
+cat     "${OUTDIR}/ordsetcover.tsv" | "${BINROOT}/setcover-ords-to-hdrs.out"  "${OUTDIR}/shuf.fasta" > "${OUTCLUST}"
+cat     "${OUTDIR}/ordsetcover.tsv" | "${BINROOT}/setcover-ords-to-fasta.out" "${OUTDIR}/shuf.fasta" > "${OUTFASTA}"           
 
-rm -r "${OUTDIR}"
+if [[ "${OUTFASTA}" != *".keep" ]]; then
+    rm -r "${OUTDIR}"
+fi
